Remove the SimplyRESTful-api package.
This package was separated to allow a client to use the interface to understand the API structure using something like Feign or MicroProfile REST Client. This turned out not to be feasible as the client would require entirely different annotations and does not allow discovery of API capabilities at runtime. At this point, there does not seem to be any benefit to keeping this separate and only introduces overhead in its maintenance.

The interfaces were merged with their corresponding interface containing the default implementation. Since those latter interfaces were the ones mainly used by the rest of the framework, this change should not break compatibility for users of the framework, except if they used constants from these removed interfaces directly. Those constants have been moved to the interface with the default implementation so the reference would need to be updated.